prevent setting empty expression, clean up expression parsing
